KinectProject: Add support to build on MacOSX
MacOS X build support has been added. However, due to the lack
of a NiTE2 distribution for this OS (I didn't download it before
PrimeSense was acquired by Apple), user tracker it's only available
when building for Windows or Linux.

+++ KinectProject/src/CoreLib/types/Point.h
@@ -113,7 +113,7 @@ public:
     explicit Point(T firstValue, Args...params)
     {
         Q_ASSERT(sizeof...(params) == N-1);
-        T data[N]={firstValue, params...};
+        T data[N]={firstValue, static_cast<float>(params)...};
         memcpy(m_data, data, N * sizeof(T));
     }
